Fetch funding wallet nonce, update tx list and simulation/submit to work with async function
